# docs/swagger.yaml
basePath: /
definitions:
...
  dto.CreateScheduleRequest:
    properties:
      blocks:
        items:
          $ref: '#/definitions/dto.TimeBlock'
        type: array
      day_of_week:
        example: 1
        type: integer
      doctor_id:
        example: uuid
        type: string
    type: object
  dto.CreateServiceRequest:
    properties:
      description:
        example: Consulta médica general
        type: string
      duration_minutes:
        example: 30
        type: integer
      name:
        example: Consulta General
        type: string
      price:
        example: 80
        type: number
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        example: Error message
        type: string
    type: object
  dto.LoginRequest:
    properties:
      email:
        example: admin@clinica.com
        type: string
      password:
        example: admin123
        type: string
    type: object
  dto.LoginResponse:
    properties:
      expires_at:
        type: string
      token:
        type: string
      user:
        $ref: '#/definitions/dto.UserResponse'
    type: object
...
  dto.ScheduleResponse:
    properties:
      blocks:
        items:
          $ref: '#/definitions/dto.TimeBlock'
        type: array
      created_at:
        type: string
      day_of_week:
        type: integer
      doctor_id:
        type: string
      id:
        type: string
    type: object
  dto.ServiceResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      duration_minutes:
        type: integer
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      price:
        type: number
    type: object
  dto.TimeBlock:
    properties:
      end_time:
        example: "13:00"
        type: string
      start_time:
        example: "09:00"
        type: string
    type: object
  dto.TimeSlot:
    properties:
      end_time:
        example: "09:30"
        type: string
      start_time:
        example: "09:00"
        type: string
    type: object
...
  dto.UserResponse:
    properties:
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      role:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: zensoftsas@gmail.com
    name: Zensoft Team
    url: https://github.com/zensoftsas/sistema-reservas
  description: |
    API completa para gestión de clínica médica con sistema de citas, servicios, horarios personalizados y analytics.

    Características principales:
    - Autenticación JWT con roles (admin, doctor, patient)
    - Sistema de citas médicas con validaciones
    - Servicios médicos con precios y duraciones
    - Horarios personalizados por doctor
    - Cálculo dinámico de slots disponibles
    - Analytics y reportes para administradores
    - Notificaciones por email (SendGrid)
    - Recordatorios automáticos
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Sistema de Reservas - Clínica Internacional API
  version: "1.0"
paths:
...
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Autenticar usuario y obtener token JWT
      parameters:
      - description: Email y password
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Login de usuario
      tags:
      - Authentication
  /api/schedules:
    post:
      consumes:
      - application/json
      description: Crear horario personalizado para un doctor (solo admin)
      parameters:
      - description: Datos del horario
        in: body
        name: schedule
        required: true
        schema:
          $ref: '#/definitions/dto.CreateScheduleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ScheduleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Crear horario de doctor
      tags:
      - Schedules
  /api/schedules/doctor/{id}:
    get:
      consumes:
      - application/json
      description: Retorna todos los horarios configurados para un doctor
      parameters:
      - description: ID del doctor
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ScheduleResponse'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Obtener horarios de un doctor
      tags:
      - Schedules
...
  /api/services/available-slots:
    get:
      consumes:
      - application/json
      description: Calcula y retorna los slots de tiempo disponibles para un doctor,
        servicio y fecha específicos
      parameters:
      - description: ID del usuario doctor
        in: query
        name: doctor_id
        required: true
        type: string
      - description: ID del servicio
        in: query
        name: service_id
        required: true
        type: string
      - description: Fecha en formato YYYY-MM-DD
        in: query
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.TimeSlot'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Obtener horarios disponibles
      tags:
      - Services
  /api/services/create:
    post:
      consumes:
      - application/json
      description: Crear un nuevo servicio médico (solo admin)
      parameters:
      - description: Datos del servicio
        in: body
        name: service
        required: true
        schema:
          $ref: '#/definitions/dto.CreateServiceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ServiceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Crear servicio
      tags:
      - Services
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...